requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1

python-dotenv~=1.0.0